[pigeon] moved the logic from bin to lib to help others wrap the command-line tools logic with their own logic.
