+++ packages/pigeon/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.24
+
+* Moved logic from bin/ to lib/ to help customers wrap up the behavior.
+
 ## 0.1.23
 
 * More Java linter and linter fixes.
