+++ packages/pigeon/lib/generator_tools.dart
@@ -8,7 +8,7 @@ import 'dart:mirrors';
 import 'ast.dart';
 
 /// The current version of pigeon. This must match the version in pubspec.yaml.
-const String pigeonVersion = '0.1.23';
+const String pigeonVersion = '0.1.24';
 
 /// Read all the content from [stdin] to a String.
 String readStdin() {
